Updated tests
Added more test classes and moved tests to the right class, based on what each test does. Also added and modified a few tests.
